# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Multiplayer Time API
# during dev, should point to your local machine
host: localhost:2000
...
basePath: /api
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
...
securityDefinitions:
  jsonWebToken:
    type: apiKey
    in: header
    name: Authorization
  jsonAdminWebToken:
    type: apiKey
    in: header
    name: Authorization
paths:
...
  /game:
    x-swagger-router-controller: game_controller
    get:
      operationId: getGames
      tags:
        - game
      #security:
      #  - jsonAdminWebToken: []
      description: get all Games
      responses:
        "200":
          description: Success
          schema:
              type: array
              items: 
                type: object
                allOf:
                  - $ref: "#/definitions/Game"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      operationId: addGame
      tags: 
        - game
      security:
        - jsonAdminWebToken: []
      summary: Add a new Game
      consumes:
        - multipart/form-data
      parameters:
        - in: formData
          name: card_image 
          type: file
          description: The file to upload.
        - in: formData
          name: logo 
          type: file
          description: The file to upload.
        - in: formData
          name: route
          type: string
        - in: formData
          name: name
          type: string 
        - in: formData
          name: description
          type: string
        - in: formData
          name: min_players
          type: integer
        - in: formData
          name: max_players
          type: integer
        - in: formData
          name: color
          type: string
        - in: formData 
          name: category_id
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ApiResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      operationId: deleteManyGames
      tags:
        - game
      #security:
      #  - jsonAdminWebToken: []
      description: remove many Games
      parameters:
        - in: body
          description: Games id's
          required: true
          name: Games
          schema:
            $ref: "#/definitions/ArrayId"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ApiResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /game/{id}:
    x-swagger-router-controller: game_controller
    get:
      operationId: getGame
      tags:
        - game
      description: get game by id
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Game"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"  
    post:
      operationId: updateGame
      tags: 
        - game
      security:
        - jsonAdminWebToken: []
      summary: Add a new Game
      consumes:
        - multipart/form-data
      parameters:
        - name: id
          type: string
          in: path
          required: true
        - in: formData
          name: card_image 
          type: file
          description: The file to upload.
        - in: formData
          name: logo 
          type: file
          description: The file to upload.
        - in: formData
          name: route
          type: string
        - in: formData
          name: name
          type: string 
        - in: formData
          name: description
          type: string
        - in: formData
          name: min_players
          type: integer
        - in: formData
          name: max_players
          type: integer
        - in: formData
          name: color
          type: string
        - in: formData 
          name: category_id
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ApiResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      operationId: deleteGame
      tags: 
        - game
      #security:
      #  - jsonAdminWebToken: []
      description: delete Game by id
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ApiResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse" 
  /game/route/{id}:
    x-swagger-router-controller: game_controller
    get:
      operationId: getGameByRoute
      tags:
        - game
      description: get game by id
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Game"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /game/recent:
    x-swagger-router-controller: game_controller
    get:
      operationId: getRecentGames
      tags:
        - game
      description: get all Games
      responses:
        "200":
          description: Success
          schema:
              type: array
              items: 
                type: object
                allOf:
                  - $ref: "#/definitions/Game"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /game/category/{id}:
    x-swagger-router-controller: game_controller
    get:
      operationId: getGamesByCategory
      tags:
        - game
      description: get all Games by Category
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
              type: array
              items: 
                type: object
                allOf:
                  - $ref: "#/definitions/Game"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"          
  ################### GAMES #######################
  ################### ROOM  #######################
  /room/invite/{id}:
    x-swagger-router-controller: room_controller
    post:
      operationId: addInvite
      tags: 
        - room
      security:
      - jsonWebToken: []
      summary: Add a new Invite
      parameters:
        - name: id
          type: string
          in: path
          required: true
        - in: body
          description: Players id's
          required: true
          name: Invites
          schema:
            $ref: "#/definitions/Invite"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ApiResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /room/invite:
    x-swagger-router-controller: room_controller
    get:
      operationId: getInvites
      tags: 
        - room
      security:
      - jsonWebToken: []
      summary: Add a new Invite
      responses:
        "200":
          description: Success
          schema:
              type: array
              items: 
                type: object
                allOf:
                  - $ref: "#/definitions/InviteInfo"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /room/invite/game/{id}:
    x-swagger-router-controller: room_controller
    get:
      operationId: getInvitesByGame
      tags: 
        - room
      security:
      - jsonWebToken: []
      summary: Add a new Invite
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
              type: array
              items: 
                type: object
                allOf:
                  - $ref: "#/definitions/InviteInfo"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /room/invite/remove/{id}:
    x-swagger-router-controller: room_controller
    get:
      operationId: removeInvite
      tags:
        - game
      description: get game by id
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ApiResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /room/invites/remove/{id}:
    x-swagger-router-controller: room_controller
    get:
      operationId: removeAllInvites
      tags:
        - game
      description: get game by id
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ApiResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  
  ################### ROOM  #######################
  ################### ADMIN #######################
...
definitions:
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  ApiResponse:
    type: object
    properties:
      success:
        type: boolean
        description: returns true if successful
      description:
        type: string
        description: a short comment 
    required:
      - success
      - description
...
  Game:
    type: object
    properties:
      id:
        type: string
      route: 
        type: string
      name: 
        type: string 
      description: 
        type: string 
      players:
        type: string
      min_players:
        type: integer
      max_players:
        type: integer
      color:
        type: string
      created_at:
        type: string
      card_image:
        type: string
      logo:
        type: string
      category:
        type: object
        properties:
          id:
            type: string
          name:
            type: string
          color:
            type: string
  ArrayId:
    type: object
    properties:
      id:
        type: array
        items:
            type: string
  Invite:
    type: object
    properties:
      game:
        type: string
      id:
        type: array
        items:
            type: string
  InviteInfo:
    type: object
    properties:
      id:
        type: string
      room:
        type: string
      game:
        type: object
        properties:
          name:
            type: string
          logo:
            type: string
          route:
            type: string
      user:
        type: object
        properties:
          username: 
            type: string
          avatar:
            type: string
...